# swagger/api.yaml
openapi: 3.0.0
info:
  title: Social Media API
  description: |
    Social Media Backend API built with NestJS and MongoDB.
    This API provides authentication, user management, posts, comments, stories, groups, pages, and messaging functionality.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register a new user
      description: Creates a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - username
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  minLength: 6
                  example: password123
                username:
                  type: string
                  example: johndoe
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: Doe
      responses:
        '201':
          description: User created successfully
        '400':
          description: Bad request
        '409':
          description: User already exists
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login user
      description: Authenticates a user and returns JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh access token
      description: Refreshes the access token using a refresh token
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
        '401':
          description: Invalid refresh token
  /auth/profile:
    get:
      tags:
        - Auth
      summary: Get user profile
      description: Returns the authenticated user's profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized

    patch:
      tags:
        - Auth
      summary: Update user profile
      description: Updates the authenticated user's profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                bio:
                  type: string
                avatar:
                  type: string
                  format: binary
      responses:
        '200':
          description: Profile updated successfully
        '401':
          description: Unauthorized
  /auth/auth-code:
    patch:
      tags:
        - Auth
      summary: Update authentication code
      description: Updates the user's authentication code
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - authCode
              properties:
                authCode:
                  type: string
      responses:
        '200':
          description: Auth code updated successfully
        '401':
          description: Unauthorized
  /auth/verify:
    get:
      tags:
        - Auth
      summary: Verify token
      description: Verifies if the current JWT token is valid
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  user:
                    type: object
        '401':
          description: Invalid token
...
  /friend-requests:
    get:
      tags:
        - Friend Requests
      summary: Get friend requests
      description: Returns friend requests for the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [sent, received]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, rejected]
      responses:
        '200':
          description: List of friend requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FriendRequest'
        '401':
          description: Unauthorized

    post:
      tags:
        - Friend Requests
      summary: Send friend request
      description: Sends a friend request to another user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - receiverId
              properties:
                receiverId:
                  type: string
      responses:
        '201':
          description: Friend request sent successfully
        '401':
          description: Unauthorized
        '400':
          description: Bad request
  /friend-requests/{id}:
    patch:
      tags:
        - Friend Requests
      summary: Respond to friend request
      description: Accepts or rejects a friend request
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [accepted, rejected]
      responses:
        '200':
          description: Friend request updated successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Friend request not found

    delete:
      tags:
        - Friend Requests
      summary: Cancel friend request
      description: Cancels a sent friend request
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Friend request cancelled successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Friend request not found
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        email:
          type: string
          format: email
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        bio:
          type: string
        avatar:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    FriendRequest:
      type: object
      properties:
        _id:
          type: string
        sender:
          $ref: '#/components/schemas/User'
        receiver:
          $ref: '#/components/schemas/User'
        status:
          type: string
          enum: [pending, accepted, rejected]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...